March Update
Project now includes some music sourced from Open Game Art.  Attribution file should be made later.  Cleaned up Block Manager by moving the map generation code into it's own class file.  Worked a bit more on the AI system.
The walking Ai is nearly ready  to be implemented in a more proper system.
I'll be slowly working on more documentation as the game expands, but it's not exactly a top priority.

This project might end fairly soon.  I think I'll plan a few more systems but a few things make me feel as if Unity isn't the best engine for a game like this, and I'm not experienced enough in the backend of 2D graphics to optimize it properly to make it work better.

+++ Assets/Scripts/Block/BlockInstance.cs
@@ -5,12 +5,12 @@ using UnityEngine;
 public class BlockInstance 
 {
     public BlockData blockData;
-    public short x;
-    public short y;
+    public int x;
+    public int y;
     public GameObject GO;
     public bool isValid = false;
 
-    public BlockInstance(BlockData d, short mx, short my) 
+    public BlockInstance(BlockData d, int mx, int my) 
     {
         blockData = d;
         x = mx;
@@ -18,7 +18,7 @@ public class BlockInstance
         isValid = true;
     }
 
-    public BlockInstance(BlockData d, short mx, short my, GameObject g)
+    public BlockInstance(BlockData d, int mx, int my, GameObject g)
     {
         blockData = d;
         x = mx;
